Fix docs regarding 'shell' param
--shell is valid only under the prepare stage. It is not used or
accepted by the run stage.

If --shell is omitted, the job will fail during the prepare step with:
gitlab-runner-kubevirt: error: missing flags: --shell=STRING

Additionally, if the --shell parameter is provided to the run step as
the docs previously suggested, once the paramter is added to the prepare
step, the run step will fail with:
gitlab-runner-kubevirt: error: unknown flag --shell, did you mean "--help"?

The driver works correctly when --shell is provided to the prepare step
and omitted from the run step.

Fixes #1